feat(structure): standardize app folders
