+++ apps/api/src/controllers/log-analysis.controller.ts
@@ -16,34 +16,38 @@ import { tmpdir } from 'node:os';
 import { join } from 'node:path';
 import type { Express } from 'express';
 
-import { ILogAnalysisService } from './ILogAnalysisService';
+import { ILogAnalysisService } from '../services/ILogAnalysisService';
 import { ParsedLog } from '@testlog-inspector/log-parser';
 import { ERR_NO_FILES } from '../common/error-messages';
 
 /**
- * POST /upload
+ * POST /analyze
  * -------------
  * Accepts one or more `.txt` or `.log` files via multipart-form-data
  *   field "files": Express.Multer.File[]
  *
  * Returns an array of ParsedLog (one result per file).
- * Dedicated controller so other upload-related actions can be added
- * without bloating `LogAnalysisController`.
  */
-@Controller('upload')
-export class UploadController {
+@Controller()
+export class LogAnalysisController {
   constructor(
     @Inject('ILogAnalysisService') private readonly service: ILogAnalysisService,
   ) {}
 
-  @Post()
+  /**
+   * Multer interceptor:
+   *   - field `files`
+   *   - max 10 files (arbitrary)
+   *   - size already limited to 50MB via global config
+   */
+  @Post('analyze')
   @HttpCode(HttpStatus.OK)
   @UseInterceptors(
     FilesInterceptor('files', 10, <MulterModuleOptions>{
       dest: join(tmpdir(), 'testlog-inspector'),
     }),
   )
-  async upload(
+  async analyze(
     @UploadedFiles() files: Express.Multer.File[],
   ): Promise<ParsedLog[]> {
     if (!files?.length) {
