+++ apps/api/src/dto/analyze-log.dto.ts
