+++ apps/api/src/dto/analyzed-result.dto.ts
