+++ apps/api/src/services/ILogAnalysisService.ts
