+++ apps/api/src/services/log-analysis.service.spec.ts
