+++ apps/api/src/services/log-analysis.service.ts
